Move AbstractBufferedEventSource to core
Moves the AbstractBufferedEventSource implementation, and its associated
tests, from the event-source-kafka module into the event-sources-core
module to make it easier to reuse in non-Kafka contexts
